refactor: move upload handling into a own controller
Will be addressing reports in Sentry.
